Bug-fixed

+++ OnlineStationary/OnlineStationary/src/main/java/com/example/OnlineStationary/module/product/dto/ProductDTO.java
@@ -1,27 +1,24 @@
 package com.example.OnlineStationary.module.product.dto;
 
+import java.math.BigDecimal;
+
 public class ProductDTO {
 
     private Long productId;
     private String name;
     private String description;
-    private Double price;
-
+    private BigDecimal price;
     private Long categoryId;
     private String categoryName;
-
     private Long brandId;
     private String brandName;
-
     private Integer stockQuantity;
 
-    public ProductDTO() {
-    }
+    public ProductDTO() {}
 
-    public ProductDTO(Long productId, String name, String description, Double price,
+    public ProductDTO(Long productId, String name, String description, BigDecimal price,
                       Long categoryId, String categoryName,
-                      Long brandId, String brandName,
-                      Integer stockQuantity) {
+                      Long brandId, String brandName, Integer stockQuantity) {
         this.productId = productId;
         this.name = name;
         this.description = description;
@@ -33,75 +30,22 @@ public class ProductDTO {
         this.stockQuantity = stockQuantity;
     }
 
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(Long productId) {
-        this.productId = productId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Double getPrice() {
-        return price;
-    }
-
-    public void setPrice(Double price) {
-        this.price = price;
-    }
-
-    public Long getCategoryId() {
-        return categoryId;
-    }
-
-    public void setCategoryId(Long categoryId) {
-        this.categoryId = categoryId;
-    }
-
-    public String getCategoryName() {
-        return categoryName;
-    }
-
-    public void setCategoryName(String categoryName) {
-        this.categoryName = categoryName;
-    }
-
-    public Long getBrandId() {
-        return brandId;
-    }
-
-    public void setBrandId(Long brandId) {
-        this.brandId = brandId;
-    }
-
-    public String getBrandName() {
-        return brandName;
-    }
-
-    public void setBrandName(String brandName) {
-        this.brandName = brandName;
-    }
-
-    public Integer getStockQuantity() {
-        return stockQuantity;
-    }
-
-    public void setStockQuantity(Integer stockQuantity) {
-        this.stockQuantity = stockQuantity;
-    }
+    public Long getProductId() { return productId; }
+    public void setProductId(Long productId) { this.productId = productId; }
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+    public String getDescription() { return description; }
+    public void setDescription(String description) { this.description = description; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
+    public Long getCategoryId() { return categoryId; }
+    public void setCategoryId(Long categoryId) { this.categoryId = categoryId; }
+    public String getCategoryName() { return categoryName; }
+    public void setCategoryName(String categoryName) { this.categoryName = categoryName; }
+    public Long getBrandId() { return brandId; }
+    public void setBrandId(Long brandId) { this.brandId = brandId; }
+    public String getBrandName() { return brandName; }
+    public void setBrandName(String brandName) { this.brandName = brandName; }
+    public Integer getStockQuantity() { return stockQuantity; }
+    public void setStockQuantity(Integer stockQuantity) { this.stockQuantity = stockQuantity; }
 }
